Remove inheritance for move form
