+++ app/models/forms/move.rb
@@ -2,6 +2,8 @@ module Forms
   class Move < Forms::Base
     REASON_WITH_DETAILS = 'other'.freeze
     COMMON_NOT_FOR_RELEASE_REASONS = %w[held_for_immigration other].freeze
+    PRISON_NOT_FOR_RELEASE_REASONS = %w[serving_sentence further_charges licence_revoke].freeze
+    POLICE_NOT_FOR_RELEASE_REASONS = %w[prison_production recall_to_prison].freeze
 
     property :to
     property :date, type: TextDate
@@ -9,6 +11,22 @@ module Forms
     optional_field :not_for_release,
       options: TOGGLE_CHOICES,
       allow_blank: false
+
+    optional_field :not_for_release_reason,
+      type: StrictString,
+      options: -> { not_for_release_reasons },
+      option_with_details: REASON_WITH_DETAILS do
+        validates :not_for_release_reason,
+          inclusion: { in: :not_for_release_reasons },
+          if: -> { not_for_release == 'yes' }
+      end
+
+    reset attributes: %i[not_for_release_reason not_for_release_reason_details],
+          if_falsey: :not_for_release
+
+    reset attributes: [:not_for_release_reason_details],
+          if_falsey: :not_for_release_reason,
+          enabled_value: REASON_WITH_DETAILS
 
     property :not_for_release_reason_details, type: StrictString
     validates :not_for_release_reason_details,
@@ -18,6 +36,14 @@ module Forms
     delegate :persisted?, to: :model
 
     validates :date, date: { not_in_the_past: true }
+
+    def not_for_release_reasons
+      if model.from_police?
+        POLICE_NOT_FOR_RELEASE_REASONS + COMMON_NOT_FOR_RELEASE_REASONS
+      else
+        PRISON_NOT_FOR_RELEASE_REASONS + COMMON_NOT_FOR_RELEASE_REASONS
+      end
+    end
 
     def not_for_release_reason_with_details
       REASON_WITH_DETAILS
@@ -50,11 +76,6 @@ module Forms
     def save
       self.to = send("to_#{to_type}")
       super
-    end
-
-    def self.form_for(move)
-      return Forms::Police::Move.new(move) if move.from_police?
-      Forms::Prison::Move.new(move)
     end
   end
 end
