+++ app/views/moves/_form.html.slim
@@ -27,7 +27,7 @@
               - else
                 = f.select "to_#{option}", option.to_s.classify.constantize.pluck(:name), {include_blank: true}, { id: option }
 
-          - Forms::Prison::Move::FREE_FORM_DESTINATION_TYPES.each do |option|
+          - Forms::Move::FREE_FORM_DESTINATION_TYPES.each do |option|
             = f.radio_concertina_option(:to_type, option) do
               label.form-label for="#{option}-text"
                 = t(option, scope: 'move.destination_types.prompts')
